+++ massa-ledger-exports/src/test_exports/mod.rs
@@ -1,0 +1,3 @@
+//! Exports testing utilities
+
+pub mod config;
