+++ massa-ledger-worker/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 
 [dependencies]
 serde_json = "1.0"
-tempfile = { version = "3.2", optional = true }    # use with testing feature
+tempfile = { version = "3.2", optional = true } # use with testing feature
 rocksdb = "0.18.0"
 nom = "7.1"
 
@@ -23,4 +23,4 @@ massa_signature = { path = "../massa-signature" }
 # for more information on what are the following features used for, see the cargo.toml at workspace level
 [features]
 instrument = ["massa_models/instrument"]
-testing = ["tempfile", "massa_models/testing"]
+testing = ["tempfile", "massa_models/testing", "massa_ledger_exports/testing"]
