+++ massa-ledger-worker/src/test_exports/mod.rs
@@ -1,4 +1,4 @@
-// exports testing utilities
+//! Exports testing utilities
 
 mod bootstrap;
 mod config;
